@@ -14,7 +14,7 @@ mixin class AuthDialogMixin {
   static const keyPadConfig = KeyPadConfig(clearOnLongPressed: true);
 
   static EdgeInsets getPadding(BuildContext context) =>
-      paddingH20 +
+      paddingHDefault +
       (ScreenSize(context) is ScreenSmall
           ? const EdgeInsets.symmetric(vertical: 12)
           : const EdgeInsets.symmetric(vertical: 32));
@@ -23,13 +23,13 @@ mixin class AuthDialogMixin {
     final theme = Theme.of(context);
     return ScreenLockConfig(
       themeData: theme,
-      backgroundColor: theme.colorScheme.surface.withOpacity(0.5),
+      backgroundColor: theme.colorScheme.surfaceTint.withOpacity(0.8),
       textStyle: theme.textTheme.titleMedium,
       titleTextStyle: theme.textTheme.titleLarge,
       buttonStyle: ButtonStyle(
-        foregroundColor: MaterialStatePropertyAll(theme.colorScheme.onSurface),
-        side: const MaterialStatePropertyAll(BorderSide.none),
-        shape: const MaterialStatePropertyAll(CircleBorder(
+        foregroundColor: WidgetStatePropertyAll(theme.colorScheme.onSurface),
+        side: const WidgetStatePropertyAll(BorderSide.none),
+        shape: const WidgetStatePropertyAll(CircleBorder(
           side: BorderSide.none,
         )),
       ),
