@@ -12,9 +12,9 @@ ScreenLockConfig getScreenLockConfig(BuildContext context) {
     textStyle: theme.textTheme.titleMedium,
     titleTextStyle: theme.textTheme.titleLarge,
     buttonStyle: ButtonStyle(
-      foregroundColor: MaterialStatePropertyAll(theme.colorScheme.onSurface),
-      side: const MaterialStatePropertyAll(BorderSide.none),
-      shape: const MaterialStatePropertyAll(CircleBorder(
+      foregroundColor: WidgetStatePropertyAll(theme.colorScheme.onSurface),
+      side: const WidgetStatePropertyAll(BorderSide.none),
+      shape: const WidgetStatePropertyAll(CircleBorder(
         side: BorderSide.none,
       )),
     ),
