@@ -1,3 +1,4 @@
+import 'package:flutter/cupertino.dart';
 import 'package:qr_flutter/qr_flutter.dart';
 
 import 'package:guardian_keyper/consts.dart';
@@ -18,8 +19,7 @@ class OnQRCodeShowDialog extends StatefulWidget {
     required String subtitle,
     required MessageModel message,
   }) =>
-      Navigator.of(context).push(MaterialPageRoute(
-        fullscreenDialog: true,
+      Navigator.of(context).push(CupertinoPageRoute(
         settings: const RouteSettings(name: route),
         builder: (_) => OnQRCodeShowDialog(
           caption: caption,
@@ -69,15 +69,20 @@ class _OnQRCodeShowDialogState extends State<OnQRCodeShowDialog> {
 
   @override
   Widget build(BuildContext context) => ScaffoldSafe(
-        header: HeaderBar(
-          caption: widget.caption,
-          rightButton: const HeaderBarButton.close(),
+        appBar: AppBar(
+          title: Text(widget.caption),
+          centerTitle: true,
+          leading: IconButton(
+            icon: const Icon(Icons.arrow_back),
+            onPressed: () {
+              Navigator.pop(context);
+            },
+          ),
         ),
         child: Column(
           children: [
             // Text
             PageTitle(
-              title: _screenSize is ScreenSmall ? null : widget.title,
               subtitle: widget.subtitle,
             ),
             // QR Code
@@ -111,7 +116,7 @@ class _OnQRCodeShowDialogState extends State<OnQRCodeShowDialog> {
             const Text('Text Code'),
             // Share Button
             Padding(
-              padding: paddingAll20,
+              padding: paddingAllDefault,
               child: Row(
                 children: [
                   Expanded(
@@ -127,7 +132,7 @@ class _OnQRCodeShowDialogState extends State<OnQRCodeShowDialog> {
                           ),
                         ),
                       ),
-                      height: buttonSize,
+                      height: kButtonSize,
                       child: Padding(
                         padding: const EdgeInsets.all(8),
                         child: Text(
@@ -148,7 +153,7 @@ class _OnQRCodeShowDialogState extends State<OnQRCodeShowDialog> {
                         ),
                       ),
                     ),
-                    height: buttonSize,
+                    height: kButtonSize,
                     child: Builder(
                       builder: (context) => IconButton(
                         icon: Transform.flip(
