@@ -14,7 +14,7 @@ class OnFailDialog extends StatelessWidget {
   Widget build(BuildContext context) => BottomSheetWidget(
         titleString: 'Invalid Code',
         textString: 'Seems like the Code you’ve just used is not valid. '
-            'Ask Guardian to share a new code.',
+            'Ask a Guardian to share a new code.',
         icon: const Icon(Icons.cancel, size: 80),
         footer: FilledButton(
           onPressed: Navigator.of(context).pop,
