@@ -14,7 +14,7 @@ class OnRejectDialog extends StatelessWidget {
   Widget build(BuildContext context) => BottomSheetWidget(
         icon: const Icon(Icons.cancel, size: 80),
         titleString: 'Request has been rejected',
-        textString: 'Guardian rejected your request to join Safe.',
+        textString: 'Guardian rejected your request to join that Safe.',
         footer: FilledButton(
           onPressed: Navigator.of(context).pop,
           child: const Text('Close'),
