@@ -19,6 +19,7 @@ class _LoadingPageState extends State<LoadingPage> {
     super.initState();
     context.read<VaultRestorePresenter>().startRequest().then(
       (message) async {
+        if (!mounted) return;
         if (message.isAccepted) {
           final wantAddAnother = await OnSuccessDialog.show(
             context,
@@ -48,54 +49,61 @@ class _LoadingPageState extends State<LoadingPage> {
   }
 
   @override
-  Widget build(BuildContext context) => Column(
-        crossAxisAlignment: CrossAxisAlignment.stretch,
-        children: [
-          // Header
-          const HeaderBar(
-            caption: 'Restore a Safe',
-            rightButton: HeaderBarButton.close(),
+  Widget build(BuildContext context) => ScaffoldSafe(
+        appBar: AppBar(
+          title: const Text('Restoring your Safe'),
+          centerTitle: true,
+          leading: IconButton(
+            icon: const Icon(Icons.arrow_back),
+            onPressed: () {
+              Navigator.pop(context);
+            },
           ),
-          // Body
-          const Padding(padding: paddingT12),
-          Padding(
-            padding: paddingAll20,
-            child: Card(
-              child: Column(
-                children: [
-                  Padding(
-                    padding: paddingT20,
-                    child: Selector<VaultRestorePresenter, bool>(
-                      selector: (_, presenter) => presenter.isWaiting,
-                      builder: (_, isWaiting, __) => Visibility(
-                        visible: isWaiting,
-                        child: const CircularProgressIndicator.adaptive(),
+        ),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.stretch,
+          children: [
+            // Body
+            const Padding(padding: paddingT12),
+            Padding(
+              padding: paddingAllDefault,
+              child: Card(
+                child: Column(
+                  children: [
+                    Padding(
+                      padding: paddingTDefault,
+                      child: Selector<VaultRestorePresenter, bool>(
+                        selector: (_, presenter) => presenter.isWaiting,
+                        builder: (_, isWaiting, __) => Visibility(
+                          visible: isWaiting,
+                          child: const CircularProgressIndicator.adaptive(),
+                        ),
                       ),
                     ),
-                  ),
-                  Padding(
-                    padding: paddingAll20,
-                    child: RichText(
-                      text: TextSpan(
-                        children: [
-                          const TextSpan(text: 'Awaiting '),
-                          TextSpan(
-                            text: context
-                                .read<VaultRestorePresenter>()
-                                .qrCode!
-                                .peerId
-                                .name,
-                            style: styleW600,
-                          ),
-                          const TextSpan(text: '’s response'),
-                        ],
+                    Padding(
+                      padding: paddingAllDefault,
+                      child: RichText(
+                        text: TextSpan(
+                          children: [
+                            const TextSpan(text: 'Awaiting '),
+                            TextSpan(
+                              text: context
+                                  .read<VaultRestorePresenter>()
+                                  .qrCode!
+                                  .peerId
+                                  .name,
+                              style: styleW600,
+                            ),
+                            const TextSpan(text: '’s response'),
+                          ],
+                        ),
                       ),
                     ),
-                  ),
-                ],
+                  ],
+                ),
               ),
             ),
-          ),
-        ],
+          ],
+        ),
       );
 }
