@@ -23,7 +23,11 @@ class OnVaultRemoveDialog extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => BottomSheetWidget(
-        icon: const Icon(Icons.warning_rounded, size: 80),
+        icon: Icon(
+          Icons.warning_rounded,
+          size: 80,
+          color: Theme.of(context).colorScheme.error,
+        ),
         titleString: 'Do you want to remove this Safe?',
         textString: 'All the Secrets from this Safe will be removed as well.',
         footer: FilledButton(
