@@ -29,22 +29,23 @@ class _GuardianWithPingTileState extends State<GuardianWithPingTile> {
           final startedAt = DateTime.now();
           final hasPong =
               await GetIt.I<VaultInteractor>().pingPeer(widget.guardian);
-          if (!mounted) return;
-          if (hasPong) {
-            final msElapsed =
-                DateTime.now().difference(startedAt).inMilliseconds;
-            showSnackBar(
-              context,
-              text: '${widget.guardian.name} is online. Ping $msElapsed ms.',
-            );
-          } else {
-            showSnackBar(
-              context,
-              text: 'Couldn’t reach out to ${widget.guardian.name}.',
-              isError: true,
-            );
+          if (context.mounted) {
+            if (hasPong) {
+              final msElapsed =
+                  DateTime.now().difference(startedAt).inMilliseconds;
+              showSnackBar(
+                context,
+                text: '${widget.guardian.name} is online. Ping $msElapsed ms.',
+              );
+            } else {
+              showSnackBar(
+                context,
+                text: 'Couldn’t reach out to ${widget.guardian.name}.',
+                isError: true,
+              );
+            }
+            setState(() => _isWaiting = false);
           }
-          setState(() => _isWaiting = false);
         },
       );
 }
