@@ -1,3 +1,4 @@
+import 'package:guardian_keyper/consts.dart';
 import 'package:guardian_keyper/app/routes.dart';
 import 'package:guardian_keyper/ui/widgets/common.dart';
 
@@ -17,32 +18,38 @@ class GuardiansExpansionTile extends StatelessWidget {
   final bool initiallyExpanded;
 
   @override
-  Widget build(BuildContext context) => ExpansionTile(
-        initiallyExpanded: initiallyExpanded,
-        childrenPadding: EdgeInsets.zero,
-        title: const Text('Safe’s Guardians'),
-        subtitle: vault.isFull
-            ? Text('${vault.size} out of ${vault.maxSize}')
-            : Row(
-                children: [
-                  const Icon(Icons.info, color: Colors.orange, size: 16),
-                  Text('  ${vault.size} out of ${vault.maxSize}'),
-                ],
+  Widget build(BuildContext context) => DecoratedBox(
+        decoration: BoxDecoration(
+          borderRadius: const BorderRadius.all(Radius.circular(kCornerRadius)),
+          color: Theme.of(context).colorScheme.surface,
+        ),
+        child: ExpansionTile(
+          initiallyExpanded: initiallyExpanded,
+          childrenPadding: EdgeInsets.zero,
+          title: const Text('Safe’s Guardians'),
+          subtitle: vault.isFull
+              ? Text('${vault.size} out of ${vault.maxSize}')
+              : Row(
+                  children: [
+                    const Icon(Icons.info, color: Colors.orange, size: 16),
+                    Text('  ${vault.size} out of ${vault.maxSize}'),
+                  ],
+                ),
+          children: [
+            for (final guardian in vault.guardians.keys)
+              guardian == vault.ownerId
+                  ? const GuardianListTile.my()
+                  : GuardianWithPingTile(guardian: guardian),
+            for (var i = vault.size; i < vault.maxSize; i++)
+              GuardianListTile.empty(
+                onTap: () => vault.isRestricted
+                    ? Navigator.of(context).pushNamed(routeVaultRestore)
+                    : Navigator.of(context).pushNamed(
+                        routeVaultGuardianAdd,
+                        arguments: vault.id,
+                      ),
               ),
-        children: [
-          for (final guardian in vault.guardians.keys)
-            guardian == vault.ownerId
-                ? const GuardianListTile.my()
-                : GuardianWithPingTile(guardian: guardian),
-          for (var i = vault.size; i < vault.maxSize; i++)
-            GuardianListTile.empty(
-              onTap: () => vault.isRestricted
-                  ? Navigator.of(context).pushNamed(routeVaultRestore)
-                  : Navigator.of(context).pushNamed(
-                      routeVaultGuardianAdd,
-                      arguments: vault.id,
-                    ),
-            ),
-        ],
+          ],
+        ),
       );
 }
