@@ -5,20 +5,21 @@ import 'package:guardian_keyper/ui/utils/screen_size.dart';
 
 export 'package:flutter/material.dart';
 
-const paddingAll20 = EdgeInsets.all(20);
-const paddingH20 = EdgeInsets.symmetric(horizontal: 20);
-const paddingV6 = EdgeInsets.symmetric(vertical: 6);
-const paddingV20 = EdgeInsets.symmetric(vertical: 20);
+const paddingAllDefault = EdgeInsets.all(kDefaultPadding);
+const paddingHDefault = EdgeInsets.symmetric(horizontal: kDefaultPadding);
+const paddingVDefault = EdgeInsets.symmetric(vertical: kDefaultPadding);
+const paddingTDefault = EdgeInsets.only(top: kDefaultPadding);
+const paddingBDefault = EdgeInsets.only(bottom: kDefaultPadding);
 const paddingT12 = EdgeInsets.only(top: 12);
-const paddingT20 = EdgeInsets.only(top: 20);
 const paddingB12 = EdgeInsets.only(bottom: 12);
-const paddingB20 = EdgeInsets.only(bottom: 20);
+const paddingV4 = EdgeInsets.symmetric(vertical: 4);
 
 const styleW600 = TextStyle(fontWeight: FontWeight.w600);
 
 class ScaffoldSafe extends StatelessWidget {
   const ScaffoldSafe({
     this.header,
+    this.appBar,
     this.drawer,
     this.child,
     this.children,
@@ -29,6 +30,7 @@ class ScaffoldSafe extends StatelessWidget {
     super.key,
   });
 
+  final PreferredSizeWidget? appBar;
   final Widget? header;
   final Widget? drawer;
   final Widget? child;
@@ -40,30 +42,25 @@ class ScaffoldSafe extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) => ColoredBox(
-        color: Theme.of(context).colorScheme.background,
+        color: Theme.of(context).colorScheme.surfaceTint,
         child: SafeArea(
           minimum: minimumPadding,
           child: Scaffold(
-            appBar: header == null
-                ? null
-                : PreferredSize(
-                    preferredSize: const Size.fromHeight(toolbarHeight),
-                    child: header!,
-                  ),
+            appBar: appBar,
             body: child ??
                 (children == null
                     ? null
                     : (isSeparated
                         ? ListView.separated(
-                            padding: paddingAll20,
+                            padding: paddingAllDefault,
                             itemCount: children!.length,
                             itemBuilder: (_, i) => children![i],
                             separatorBuilder: (_, __) => isSeparatorSmall
                                 ? const Padding(padding: paddingT12)
-                                : const Padding(padding: paddingT20),
+                                : const Padding(padding: paddingTDefault),
                           )
                         : ListView(
-                            padding: paddingAll20,
+                            padding: paddingAllDefault,
                             children: children!,
                           ))),
             bottomNavigationBar: bottomNavigationBar,
@@ -75,90 +72,6 @@ class ScaffoldSafe extends StatelessWidget {
       );
 }
 
-class HeaderBar extends StatelessWidget {
-  const HeaderBar({
-    super.key,
-    this.caption = '',
-    this.leftButton,
-    this.rightButton,
-    this.isTransparent = false,
-  });
-
-  final String caption;
-  final Widget? leftButton;
-  final Widget? rightButton;
-  final bool isTransparent;
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    return Container(
-      height: toolbarHeight,
-      color: isTransparent ? Colors.transparent : theme.colorScheme.background,
-      child: Row(
-        crossAxisAlignment: CrossAxisAlignment.center,
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          SizedBox.square(
-            dimension: toolbarHeight,
-            child: leftButton,
-          ),
-          Text(
-            caption,
-            maxLines: 1,
-            softWrap: false,
-            overflow: TextOverflow.ellipsis,
-            style: theme.textTheme.titleMedium,
-          ),
-          SizedBox.square(
-            dimension: toolbarHeight,
-            child: rightButton,
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class HeaderBarButton extends StatelessWidget {
-  const HeaderBarButton({
-    required this.icon,
-    this.onPressed,
-    super.key,
-  });
-
-  const HeaderBarButton.close({
-    this.onPressed,
-    super.key,
-  }) : icon = Icons.close;
-
-  const HeaderBarButton.back({
-    this.onPressed,
-    super.key,
-  }) : icon = Icons.arrow_back;
-
-  const HeaderBarButton.more({
-    required this.onPressed,
-    super.key,
-  }) : icon = Icons.keyboard_control;
-
-  final IconData icon;
-  final VoidCallback? onPressed;
-
-  @override
-  Widget build(BuildContext context) => Container(
-        decoration: BoxDecoration(
-          color: Theme.of(context).colorScheme.secondary,
-          shape: BoxShape.circle,
-        ),
-        margin: const EdgeInsets.all(14),
-        child: IconButton(
-          onPressed: onPressed ?? Navigator.of(context).pop,
-          icon: Icon(icon),
-        ),
-      );
-}
-
 class PageTitle extends StatelessWidget {
   const PageTitle({
     super.key,
@@ -166,23 +79,26 @@ class PageTitle extends StatelessWidget {
     this.title,
     this.subtitle,
     this.subtitleSpans,
+    this.color,
   });
 
   final Widget? icon;
   final String? title;
   final String? subtitle;
   final List<TextSpan>? subtitleSpans;
+  final Color? color;
 
   @override
   Widget build(BuildContext context) {
     final paddingTop = switch (ScreenSize(context)) {
       ScreenSmall _ => paddingT12,
-      ScreenMedium _ => paddingT20,
+      ScreenMedium _ => paddingTDefault,
       _ => const EdgeInsets.only(top: 32),
     };
     final theme = Theme.of(context);
+    final textColor = color ?? theme.colorScheme.onSurface;
     return Padding(
-      padding: paddingH20,
+      padding: paddingHDefault,
       child: Column(
         mainAxisSize: MainAxisSize.min,
         children: [
@@ -200,23 +116,24 @@ class PageTitle extends StatelessWidget {
                 textAlign: TextAlign.center,
                 text: TextSpan(
                   text: title,
-                  style: theme.textTheme.titleLarge,
+                  style: theme.textTheme.titleLarge!.copyWith(color: textColor),
                 ),
               ),
             ),
           // Subtitle
           if (subtitle != null || subtitleSpans != null)
             Padding(
-              padding: paddingT20,
+              padding: paddingTDefault,
               child: RichText(
                 textAlign: TextAlign.center,
                 text: TextSpan(
                   text: subtitle,
                   children: subtitleSpans,
-                  style: const TextStyle(
+                  style: TextStyle(
                     height: 1.5,
                     fontSize: 16,
                     fontWeight: FontWeight.w400,
+                    color: textColor,
                   ),
                 ),
               ),
@@ -250,7 +167,7 @@ class BottomSheetWidget extends StatelessWidget {
   Widget build(BuildContext context) {
     final textTheme = Theme.of(context).textTheme;
     return Padding(
-      padding: paddingAll20,
+      padding: paddingAllDefault,
       child: Column(
         mainAxisSize: MainAxisSize.min,
         crossAxisAlignment: CrossAxisAlignment.stretch,
@@ -285,10 +202,10 @@ class BottomSheetWidget extends StatelessWidget {
               ),
             ),
           // Body
-          if (body != null) Padding(padding: paddingT20, child: body),
+          if (body != null) Padding(padding: paddingTDefault, child: body),
           // Footer
-          if (footer != null) Padding(padding: paddingT20, child: footer),
-          const Padding(padding: paddingB20),
+          if (footer != null) Padding(padding: paddingTDefault, child: footer),
+          const Padding(padding: paddingBDefault),
         ],
       ),
     );
@@ -299,7 +216,7 @@ ScaffoldFeatureController<SnackBar, SnackBarClosedReason> showSnackBar(
   BuildContext context, {
   String? text,
   List<TextSpan>? textSpans,
-  Duration duration = snackBarDuration,
+  Duration duration = kSnackBarDuration,
   bool isFloating = false,
   bool isError = false,
 }) {
@@ -307,7 +224,7 @@ ScaffoldFeatureController<SnackBar, SnackBarClosedReason> showSnackBar(
   return ScaffoldMessenger.of(context).showSnackBar(SnackBar(
     duration: duration,
     behavior: isFloating ? SnackBarBehavior.floating : null,
-    margin: paddingAll20,
+    margin: paddingAllDefault,
     backgroundColor:
         isError ? theme.colorScheme.error : theme.snackBarTheme.backgroundColor,
     content: RichText(
